fix: resolve build errors due to Media3 bump and type mismatch

+++ app/src/main/kotlin/com/music/echo/yandeximport/YandexImportRepository.kt
@@ -159,7 +159,7 @@ class YandexImportRepository @Inject constructor(
                 SpotifyMapper.matchScore(
                     spotifyTitle = track.title,
                     spotifyArtist = track.artists,
-                    spotifyDurationMs = track.durationMs,
+                    spotifyDurationMs = track.durationMs.toInt(),
                     candidateTitle = candidate.title,
                     candidateArtist = candidate.artists.joinToString(" ") { it.name },
                     candidateDurationSec = candidate.duration,
